+++ src/main/java/org/dx42/needy/report/ArtifactsWithMultipleVersionsHtmlReport.groovy
@@ -40,6 +40,7 @@ class ArtifactsWithMultipleVersionsHtmlReport extends AbstractHtmlReport {
 					unescaped << notesHtml
 				}
 				out << buildDependencyTable(sortedMap)
+                out << buildApplicationList(dependencies)
 			}
 		}
 	}